Fix and test cum-mean-covs
