+++ posteriors/__init__.py
@@ -20,6 +20,7 @@ from posteriors.utils import per_samplify
 from posteriors.utils import is_scalar
 from posteriors.utils import L_from_flat
 from posteriors.utils import L_to_flat
+from posteriors.utils import cumulative_mean_and_cov
 
 from posteriors.tree_utils import tree_size
 from posteriors.tree_utils import tree_extract
